Changes to be committed: 	modified:   nuttx/arch/risc-v/src/sg2002/Make.defs 	modified:   nuttx/arch/risc-v/src/sg2002/sg2002_gpio.c 	modified:   nuttx/arch/risc-v/src/sg2002/sg2002_start.c 	modified:   nuttx/drivers/mailbox/mailbox_driver.c
1. gpio driver update

+++ nuttx/arch/risc-v/src/sg2002/sg2002_start.c
@@ -80,7 +80,7 @@ void sg2002_clear_bss(void)
 {
 	uint32_t *dest = NULL;
 
-	for (dest = (uint32_t *)_sbss; dest < (uint32_t *)_ebss; )
+	for (dest = (uint32_t *)((uintptr_t)_sbss); dest < (uint32_t *)((uintptr_t)_ebss);)
     {
     	*dest++ = 0;
     }
